modified:   src/app/services/login.service.ts

+++ src/app/services/login.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { BehaviorSubject, Observable, of } from 'rxjs';
-import { catchError, tap } from 'rxjs/operators';
+import { BehaviorSubject, Observable, of, timer } from 'rxjs';
+import { catchError, tap, switchMap } from 'rxjs/operators';
 import { environment } from 'src/environments/environment';
 import { Router } from '@angular/router';
 import { CookieService } from 'ngx-cookie-service';
@@ -11,78 +11,82 @@ import { ToastrService } from 'ngx-toastr';
   providedIn: 'root',
 })
 export class LoginService {
-  // private backendUrl = 'http://192.168.0.8:3000/api/users/';
-  private backendUrl = `${environment.apiUrl}users/`
+  private backendUrl = `${environment.apiUrl}users/`;
   private isLoggedInSubject: BehaviorSubject<boolean> =
     new BehaviorSubject<boolean>(false);
   public isLoggedIn$: Observable<boolean> =
     this.isLoggedInSubject.asObservable();
-  private username: string = '';
-
   private readonly COOKIE_Name = 'username';
+  private sessionTimeoutMinutes = 60;
+
+  private sessionTimeoutTimer$: Observable<number>;
 
   constructor(
     private http: HttpClient,
     private router: Router,
     private cookieService: CookieService,
     private toastr: ToastrService
-  ) {}
+  ) {
+    this.checkLoggedIn();
+
+    this.sessionTimeoutTimer$ = this.isLoggedIn$.pipe(
+      switchMap((isLoggedIn) => {
+        if(isLoggedIn) {
+          return timer(this.sessionTimeoutMinutes * 60 * 1000);
+        } else {
+          return of(null);
+        }
+      })
+    );
+
+    this.sessionTimeoutTimer$.subscribe(() => {
+      this.logoutDueToInactivity();
+    })
+  };
 
   onLogin(username: string, password: string) {
-
     if (!username || !password) {
       this.toastr.warning('Ingrese usuario y contraseña');
-    } else {
-      const loginData = {
-        username: username,
-        password: password,
-      };
-      this.http.get(`${this.backendUrl}${username}`).subscribe({
-        next: (response: any) => {
-          // console.log(response.details);
-          if (response.message) {
-            const headers = new HttpHeaders({
-              'Content-Type': 'application/json',
-              Authorization: 'Bearer ' + environment.SECRET_KEY,
-            });
-            this.http
-              .post<any>(`${this.backendUrl}login/`, loginData, { headers })
-              .pipe(
-                tap((response) => {
-                  if (response.success === true) {
-                    localStorage.setItem('token', response.message.token);
-                    const name = response.message.name;
-                    this.cookieService.set(this.COOKIE_Name, name);
-                    this.setLoggedIn(true);
-                    this.toastr.success(
-                      `¡Bienvenido ${this.capitalizeFirstLetter(name)}!`
-                    );
-                    setTimeout(() => {
-                      this.router.navigate(['/home']);
-                    }, 500);
-                  } else {
-                    this.setLoggedIn(false);
-                  }
-                }),
-                catchError((error) => {
-                  this.toastr.error('Credenciales inválidas');
-                  return of(null);
-                })
-              )
-              .subscribe();
-          } else {
-            this.toastr.warning('Error consultando el usuario');
-          }
-        },
-        error: (error: any) => {
-          this.toastr.error(`Usuario ${this.capitalizeFirstLetter(username)} no registrado`)
-        }
-      });
+      return;
     }
-    
+
+    const loginData = {
+      username: username,
+      password: password,
+    };
+
+    this.http
+      .post<any>(`${this.backendUrl}login/`, loginData)
+      .pipe(
+        tap((response) => {
+          if (response.success === true) {
+            const name = response.message.name;
+            localStorage.setItem('token', response.message.token);
+            this.cookieService.set(this.COOKIE_Name, name);
+            this.setLoggedIn(true);
+            this.toastr.success(
+              `¡Bienvenido ${this.capitalizeFirstLetter(name)}!`
+            );
+            setTimeout(() => {
+              this.router.navigate(['/home']);
+            }, 500);
+          } else {
+            this.setLoggedIn(false);
+            this.toastr.error('Credenciales inválidas');
+          }
+        }),
+        catchError((error) => {
+          this.setLoggedIn(false);
+          this.toastr.error(
+            `Usuario ${this.capitalizeFirstLetter(username)} no registrado`
+          );
+          return of(null);
+        })
+      )
+      .subscribe();
   }
 
-  capitalizeFirstLetter(value: string): string {
+  private capitalizeFirstLetter(value: string): string {
     return value.charAt(0).toUpperCase() + value.slice(1);
   }
 
@@ -91,23 +95,26 @@ export class LoginService {
   }
 
   onLogout() {
+    const token = localStorage.getItem('token');
+    if (!token) {
+      this.toastr.error('No se ha iniciado sesión');
+      return;
+    }
+
     const headers = new HttpHeaders({
       'Content-Type': 'application/json',
-      Authorization: 'Bearer ' + localStorage.getItem('token'),
+      Authorization: 'Bearer ' + token,
     });
+
     this.http
-      .post<any>(
-        `${this.backendUrl}logout/`,
-        { token: localStorage.getItem('token') },
-        { headers }
-      )
+      .post<any>(`${this.backendUrl}logout/`, { token: token }, { headers })
       .pipe(
         tap(() => {
           localStorage.removeItem('token');
           this.setLoggedIn(false);
           this.cookieService.delete(this.COOKIE_Name);
-          this.username = '';
-          this.toastr.success('Sesión cerrada exitosamente')
+          this.toastr.success('Sesión cerrada exitosamente');
+          this.router.navigate(['/login']);
         }),
         catchError((error) => {
           this.toastr.error('Error al cerrar sesión');
@@ -115,9 +122,7 @@ export class LoginService {
           return of(null);
         })
       )
-      .subscribe(() => {
-        this.router.navigate(['/login']);
-      });
+      .subscribe();
   }
 
   isLoggedIn(): Observable<boolean> {
@@ -134,10 +139,11 @@ export class LoginService {
 
   checkLoggedIn(): void {
     const token = localStorage.getItem('token');
-    if (token) {
-      this.setLoggedIn(true);
-    } else {
-      this.setLoggedIn(false);
-    }
+    this.setLoggedIn(!!token);
+  }
+
+  private logoutDueToInactivity() {
+    this.toastr.warning("Sesión cerrada por inactividad");
+    this.onLogout();
   }
 }
